test: make Nightly prerequisites deterministic
